Removed the "shared" folder
Moved the contents of "shared" folder to "telemetryui" and deleted it.
